requirements.txt
flask==2.3.3
qrcode==7.4.2
pillow==10.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
gunicorn==21.2.0